pyproject.toml
[tool.poetry]
name = "fastapi-cerbos"
version = "0.1.0"
description = ""
authors = ["Rohit Ghumare <ghumare64@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
uvicorn = "^0.29.0"
pymysql = "^1.1.0"
cerbos = "^0.10.5"
cryptography = "^42.0.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
